+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v0.2.0 (10/31/2022)
+
+### Feature
+
+- Added train/test splits for NW-UCLA dataset
+- Added cross-setup evaluation for NTU 120 dataset
+
+### Fix
+
+- Make sure path isn't `None` in NTU dataset when `num_classes=120`
+
 ## v0.1.0 (28/10/2022)
 
 - First release of `torch_skeleton`